pyproject.toml
[tool.pdm]

[project]
name = "saberbld"
version = "0.1.0"
description = "CLI application for comfort view information about builds ant tasks"
authors = [{ name = "Andrew Laptev", email = "nickname.avast@gmail.com" }]
dependencies = ["typer[all]~=0.9.0", "pyyaml~=6.0", "pydantic~=1.10.7"]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[project.optional-dependencies]
tests = ["pytest~=7.3.1", "pytest-cov~=4.0.0"]

[project.scripts]
saberbld = "saberbld.main:app"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
